Moved VisibleScreenService to its own package
