+++ android-agent/src/main/java/io/opentelemetry/android/features/screenattrs/ScreenAttributesSpanProcessor.java
@@ -7,7 +7,7 @@ package io.opentelemetry.android.features.screenattrs;
 
 import static io.opentelemetry.android.common.RumConstants.SCREEN_NAME_KEY;
 
-import io.opentelemetry.android.internal.services.VisibleScreenService;
+import io.opentelemetry.android.internal.services.visiblescreen.VisibleScreenService;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.trace.ReadWriteSpan;
 import io.opentelemetry.sdk.trace.ReadableSpan;
